+++ Renderer/VulkanRenderer.cpp
@@ -761,7 +761,7 @@ void VulkanRenderer::createVertexBuffer() {
 
 	unsigned char* stagingData = reinterpret_cast<unsigned char*>(this->allocator.mapMemory(SBAllocation));
 	size_t vbOffset = 0, ibOffset = 0;
-	size_t vertexCount = 0, indexCount = 0;
+	uint32_t vertexCount = 0, indexCount = 0;
 	for (auto& mesh : meshes) {
 		size_t len = mesh->vertices.size() * sizeof(Vertex);
 		memcpy(stagingData + vbOffset, mesh->vertices.data(), len);
@@ -777,8 +777,8 @@ void VulkanRenderer::createVertexBuffer() {
 		
 		mesh->firstIndexOffset = indexCount;
 		mesh->firstVertexOffset = vertexCount;
-		indexCount += mesh->indices.size();
-		vertexCount += mesh->vertices.size();
+		indexCount += static_cast<uint32_t>(mesh->indices.size());
+		vertexCount += static_cast<uint32_t>(mesh->vertices.size());
 	}
 	this->allocator.unmapMemory(SBAllocation);
 
@@ -858,9 +858,9 @@ void VulkanRenderer::drawMeshes(const std::vector<std::shared_ptr<Mesh>>& meshes
 		cb.pushConstants<glm::mat4x4>(this->pipelineLayout, vk::ShaderStageFlagBits::eVertex, 0, pushConstants);
 
 		if (mesh->isIndexed)
-			cb.drawIndexed(mesh->indices.size(), 1, mesh->firstIndexOffset, 0, 0);
+			cb.drawIndexed(static_cast<uint32_t>(mesh->indices.size()), 1, mesh->firstIndexOffset, 0, 0);
 		else
-			cb.draw(mesh->vertices.size(), 1, mesh->firstVertexOffset, 0);
+			cb.draw(static_cast<uint32_t>(mesh->vertices.size()), 1, mesh->firstVertexOffset, 0);
 	}
 }
